move pages to components for new register page
