+++ client/src/components/index.ts
@@ -6,7 +6,10 @@ export { default as CustomButton } from './CustomButton/CustomButton';
 export { default as ResetPassword } from './ResetPassword/ResetPassword';
 export { default as Header } from './Header/Header';
 export { default as Map } from './Map/Map';
-export { default as LocationNotification } from './LocationNotification/LocationNotification'
+export { default as LocationNotification } from './LocationNotification/LocationNotification';
 export { default as ResetPasswordVerify } from './ResetPasswordVerify/ResetPasswordVerify';
 export { default as CommunityCard } from './CommunityCard/CommunityCard';
 export { default as EnableNotification } from './EnableNotification/EnableNotification';
+export { default as CreateAccount } from './CreateAccount/CreateAccount';
+export { default as SignUp } from './SignUp/SignUp';
+export { default as GeneralInfo } from './SignUp/GeneralInfo/GeneralInfo';
