+++ client/src/main.tsx
@@ -3,17 +3,7 @@ import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { store } from './store';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import {
-  Chat,
-  Community,
-  Home,
-  Login,
-  Profile,
-  SignUp,
-  GeneralInfo,
-  CreateAccount,
-  Instructions,
-} from '@/pages';
+import { Chat, Community, Home, Login, Profile, Instructions, Register } from '@/pages';
 import { PrivateRoute } from '@/utils';
 
 // Collection of paths for the site.
@@ -39,24 +29,16 @@ const router = createBrowserRouter([
     element: <Login />,
   },
   {
-    path: '/sign-up',
-    element: <SignUp />,
-  },
-  {
     path: '/home', // for testing only
     element: <Home />,
   },
   {
-    path: '/create-account',
-    element: <CreateAccount />,
-  },
-  {
-    path: '/general-info',
-    element: <GeneralInfo />,
-  },
-  {
     path: '/instructions',
     element: <Instructions />,
+  },
+  {
+    path: '/register',
+    element: <Register />,
   },
 ]);
