+++ client/src/pages/Login/Login.tsx
@@ -7,10 +7,10 @@ import styles from './Login.module.scss';
 import { useLoginMutation } from '@/store';
 import { useNavigate } from 'react-router-dom';
 import { TextField, PasswordField, CustomButton } from '@/components';
-import logo from '../../assets/Updated_RoadWatch_Logo.svg'; 
+import logo from '../../assets/Updated_RoadWatch_Logo.svg';
 
 export default function Login(): JSX.Element {
-   // State management for user input and password
+  // State management for user input and password
   const [userInput, setUserInput] = useState('');
   const [password, setPassword] = useState('');
 
@@ -35,7 +35,6 @@ export default function Login(): JSX.Element {
         }
       }
     } catch (err) {
-
       // Log any errors to the console
       console.log(err);
     }
@@ -43,26 +42,31 @@ export default function Login(): JSX.Element {
 
   // Render login form
   return (
-      <div className={styles['Login__LoginContainer']}>
-        <div className={styles['Login__whiteContainer']}>
-          <div className={styles['Login__logo']}> 
-            <img src = {logo} alt = {"Roadwatch Logo"}/>
-          </div>
-          <TextField header="Username or Email" setInputValue={setUserInput} />
-          <PasswordField header="Password" setInputValue={setPassword} />
-          <div className={styles['Login__textCentered']}>
-            <CustomButton onClick={handleLogin}>Login</CustomButton>
-            <br /><br />
-            <b>Forgot Password</b>
-            <br /><br />
-            <div className={styles['Login__registerHereContainer']}>
-                <p> {"Don't have an account?"}
-                {/* Once 'register page' is completed update to <a> element to create a hyperlink.
-                    ex. <a href="/register" className={styles.Login__registerHereBtn}>Register here</a> */}
-                <span className={styles['Login__registerHereBtn']}> Register here</span></p> 
-              </div>
+    <div className={styles['Login__LoginContainer']}>
+      <div className={styles['Login__whiteContainer']}>
+        <div className={styles['Login__logo']}>
+          <img src={logo} alt={'Roadwatch Logo'} />
+        </div>
+        <TextField header="Username or Email" setInputValue={setUserInput} />
+        <PasswordField header="Password" setInputValue={setPassword} />
+        <div className={styles['Login__textCentered']}>
+          <CustomButton onClick={handleLogin}>Login</CustomButton>
+          <br />
+          <br />
+          <b>Forgot Password</b>
+          <br />
+          <br />
+          <div className={styles['Login__registerHereContainer']}>
+            <p>
+              {' '}
+              {"Don't have an account? "}
+              <a href="/register" className={styles.Login__registerHereBtn}>
+                Register here
+              </a>
+            </p>
           </div>
         </div>
       </div>
+    </div>
   );
 }
