+++ client/src/pages/Register/Register.module.scss
@@ -1,0 +1,3 @@
+.Register {
+
+}
