+++ client/src/pages/Register/Register.tsx
@@ -1,0 +1,6 @@
+import styles from './Register.module.scss';
+import { GeneralInfo, SignUp, CreateAccount } from '@/components';
+
+export default function Register(): JSX.Element {
+  return <div className={styles['Register']}>Register</div>;
+}
