+++ client/src/pages/index.ts
@@ -13,14 +13,8 @@ export * from './Login/Login';
 export { default as Profile } from './Profile/Profile';
 export * from './Profile/Profile';
 
-export { default as SignUp } from './SignUp/SignUp';
-export * from './SignUp/SignUp';
-
-export { default as CreateAccount } from './CreateAccount/CreateAccount';
-export * from './CreateAccount/CreateAccount';
-
 export { default as Instructions } from './Instructions/Instructions';
 export * from './Instructions/Instructions';
 
-export { default as GeneralInfo } from './SignUp/GeneralInfo/GeneralInfo';
-export * from './SignUp/GeneralInfo/GeneralInfo';
+export { default as Register } from './Register/Register';
+export * from './Register/Register';
